+++ tests/test.h
@@ -3,12 +3,15 @@
 
 #include <stdbool.h>
 #include <stddef.h>
+#include <sys/stat.h>
 
 void test_run(void (*fn)());
 void test_fail();
 void assert_true(bool);
 
 void assert_string_equal(char *, char *);
+
+void assert_mode_equal(mode_t, mode_t);
 
 #endif